pyproject.toml
[tool.poetry]
name = "schafkopf-scheduler"
version = "0.1.0"
description = ""
authors = ["Lukas Bernhard <strange85@web.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
boto3 = "^1.35.49"
fastapi = "^0.115.4"
mangum = "^0.19.0"
uvicorn = "^0.32.0"
moto = {extras = ["dynamodb"], version = "^5.0.25"}
pytest = "^8.3.4"
responses = "^0.25.3"
freezegun = "^1.5.1"
httpx = "^0.28.1"
pandas = "^2.2.3"
black = "^25.1.0"
pre-commit = "^4.2.0"

[tool.poetry.group.beachbooker.dependencies]
selenium = "^4.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
